Refactor: separate a binaryImageProcess util from takePhoto
