+++ src/utils/camera/takePhoto.js
@@ -18,22 +18,23 @@ export const handleTakePhoto = async (cameraRef, flash) => {
   await CameraRoll.saveAsset(`file://${photo.path}`, {
     type: 'photo',
   });
+  return photo.path;
 
-  const base64jpg = await RNFS.readFile(photo.path, 'base64');
-  const srcMat = OpenCV.base64ToMat(base64jpg);
-  const rotatedMat = OpenCV.createObject(
-    ObjectType.Mat,
-    0,
-    0,
-    DataTypes.CV_8UC4,
-  );
-  OpenCV.invoke('rotate', srcMat, rotatedMat, 0);
-  const dstMat = OpenCV.createObject(ObjectType.Mat, 0, 0, DataTypes.CV_8U);
-  // OpenCV.invoke('Canny', rotatedMat, dstMat, 80, 120);
-  OpenCV.invoke('Canny', rotatedMat, dstMat, 120, 180);
+  // const base64jpg = await RNFS.readFile(photo.path, 'base64');
+  // const srcMat = OpenCV.base64ToMat(base64jpg);
+  // const rotatedMat = OpenCV.createObject(
+  //   ObjectType.Mat,
+  //   0,
+  //   0,
+  //   DataTypes.CV_8UC4,
+  // );
+  // OpenCV.invoke('rotate', srcMat, rotatedMat, 0);
+  // const dstMat = OpenCV.createObject(ObjectType.Mat, 0, 0, DataTypes.CV_8U);
+  // // OpenCV.invoke('Canny', rotatedMat, dstMat, 80, 120);
+  // OpenCV.invoke('Canny', rotatedMat, dstMat, 120, 180);
 
-  const result = OpenCV.toJSValue(dstMat);
-  OpenCV.clearBuffers();
+  // const result = OpenCV.toJSValue(dstMat);
+  // OpenCV.clearBuffers();
 
-  return result.base64;
+  // return result.base64;
 };
